+++ storage/junctions/json/json-writer.js
@@ -1,5 +1,6 @@
 "use strict";
 
+const Storage = require('../../storage');
 const { StorageWriter } = require('../storage-junction');
 const { StorageError } = require('../../types');
 const { logger } = require('@dictadata/lib');
@@ -78,7 +79,7 @@ module.exports = exports = class JSONWriter extends StorageWriter {
 
     try {
       // open file stream
-      this.stfs = await this.junction.getFileSystem();
+      this.stfs = await Storage.activateFileSystem(this.junction.smt, this.junction.options);
       this.ws = await this.stfs.createWriteStream(this.options);
 
       this.ws.on('error', (err) => {
@@ -200,6 +201,9 @@ module.exports = exports = class JSONWriter extends StorageWriter {
         if (this.ws.fs_ws_promise)
           await this.ws.fs_ws_promise;
       }
+
+      if (this.stfs)
+        this.stfs.relax();
 
       callback();
     }