+++ storage/junctions/rest/rest-junction.js
@@ -1,5 +1,6 @@
 "use strict";
 
+const Storage = require('../../storage');
 const StorageJunction = require('../storage-junction');
 const { StorageResults, StorageError } = require('../../types');
 const { logger } = require('@dictadata/lib');
@@ -66,7 +67,7 @@ class RESTJunction extends StorageJunction {
           logger.warn("rest codify reader: " + error.message);
         });
 
-        let codify = await this.createTransform("codify", options);
+        let codify = await Storage.activateTransform("codify", options);
         await pipeline(reader, codify);
 
         let encoding = codify.encoding;